third_party/WebKit/LayoutTests/http/tests/security/vibrate_in_cross_origin_iframe_blocked-expected.txt

CONSOLE ERROR: line 10: A call of navigator.vibrate will be no-op inside cross-origin iframes: https://www.chromestatus.com/feature/5682658461876224.
